Update KafkaSink.java
CPU occupancy rate rather than 80%, even if has not events to process.
if return BACKOFF status when has not events to process, it need to sleep 1s~5s . so I would like to define a BACKOFF_SLEEP_MILLISECOND to control the short time sleep.
In my company, I use this change and custom selector (not use regex) to run. CPU occupancy rate is lower 5%
thank you!